requirements.txt
python-dotenv == 1.1.1
paho-mqtt == 2.1.0
requests == 2.32.5
cryptography == 45.0.6